Codegen/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.1.1
pymongo==4.3.3
pydantic==2.7.0
python-dotenv==1.0.1
